move function to `adapter-utils.js`
