+++ packages/kit/rollup.config.js
@@ -49,7 +49,7 @@ export default [
 			ssr: 'src/runtime/server/index.js',
 			node: 'src/core/node/index.js',
 			'install-fetch': 'src/install-fetch.js',
-			utils: 'src/core/utils.js'
+			'adapter-utils': 'src/core/adapter-utils.js'
 		},
 		output: {
 			dir: 'dist',
